Finish part 2
I had to make this more efficient; trying to run even the example for 256
days with the old implementation was taking a huge amount of time. All
the vectors and copying, I imagine.

When first reading through the problem, I had the idea to use maps with
just counts. But I went for the simpler solution first (that matched the
example in the problem) and optimized later.